+++ CHANGELOG.md
@@ -13,6 +13,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Changed
 
 - Client-side validation of `short_description` for visualisation catalogue items no longer occurs.
+- The ability to see what Datasets a visualisation has access to
 
 ## 2020-04-22